+++ assets/music/default.txt
@@ -1,0 +1,1 @@
+240 50 240 50 240 50 240 50 240 50 240 50
